Feature/server/ai

+++ server/src/modules/player/player.h
@@ -31,12 +31,12 @@ typedef enum player_rotation_e {
  * @brief Player structure.
  * A player (also called a client ; or a Trantorian, an inhabitant of Trantor,
  * the planet where the game takes place) is an entity on the map,
- * which can perform actions like moving, taking objects....
+ * which can perform actions like moving, taking resources....
  */
 typedef struct player_s {
     /** @brief The amount of time (in time units)
      * the player can still stay alive. If it reaches 0, the player dies */
-    u_int64_t life_span;
+    double life_span; // TODO: init to 1260
 
     /** @brief The player's x position on the map */
     u_int64_t x;
